requirements.txt
requests
schedule
python-dotenv
openai
beautifulsoup4
google-genai
feedparser
yfinance
streamlit
plotly
python-json-logger
pytest
pytest-mock
psycopg2-binary
fastapi
uvicorn
simhash
dateparser
pytz
pandas
GoogleNews
sentence-transformers
scikit-learn
torch
socksio
